Move ThreadUtils to standalone module.
