+++ app/src/main/java/org/mozilla/rocket/privately/PrivateMode.kt
@@ -11,8 +11,8 @@ import android.preference.PreferenceManager
 import org.mozilla.focus.utils.AppConfigWrapper
 import org.mozilla.focus.utils.AppConstants
 import org.mozilla.focus.utils.FileUtils
-import org.mozilla.focus.utils.ThreadUtils
 import org.mozilla.rocket.component.PrivateSessionNotificationService
+import org.mozilla.threadutils.ThreadUtils
 import java.io.File
