+++ components/utils/threadutils/src/main/java/org/mozilla/threadutils/ThreadUtils.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.focus.utils;
+package org.mozilla.threadutils;
 
 import android.os.Handler;
 import android.os.Looper;
